+++ data/provider/src/main/java/tech/pegasys/teku/api/ChainDataProvider.java
@@ -73,6 +73,7 @@ import tech.pegasys.teku.spec.datastructures.state.CommitteeAssignment;
 import tech.pegasys.teku.spec.datastructures.state.SyncCommittee;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
 import tech.pegasys.teku.spec.datastructures.state.versions.electra.PendingDeposit;
+import tech.pegasys.teku.spec.datastructures.state.versions.electra.PendingPartialWithdrawal;
 import tech.pegasys.teku.spec.logic.common.statetransition.epoch.status.ValidatorStatuses;
 import tech.pegasys.teku.spec.logic.common.statetransition.exceptions.EpochProcessingException;
 import tech.pegasys.teku.spec.logic.common.statetransition.exceptions.SlotProcessingException;
@@ -759,27 +760,59 @@ public class ChainDataProvider {
 
   private Optional<ObjectAndMetaData<SszList<PendingDeposit>>> getPendingDeposits(
       final Optional<StateAndMetaData> maybeStateAndMetadata) {
-    if (maybeStateAndMetadata.isPresent()) {
-      if (!maybeStateAndMetadata
-          .get()
-          .getMilestone()
-          .isGreaterThanOrEqualTo(SpecMilestone.ELECTRA)) {
-        throw new BadRequestException(
-            "The state was successfully retrieved, but was prior to electra and does not contain pending deposits.");
-      }
-      return maybeStateAndMetadata.map(
-          stateAndMetaData -> {
-            final SszList<PendingDeposit> deposits =
-                stateAndMetaData.getData().toVersionElectra().orElseThrow().getPendingDeposits();
-            ;
-            return new ObjectAndMetaData<>(
-                deposits,
-                stateAndMetaData.getMilestone(),
-                stateAndMetaData.isExecutionOptimistic(),
-                stateAndMetaData.isCanonical(),
-                stateAndMetaData.isFinalized());
-          });
+
+    checkMinimumMilestone(maybeStateAndMetadata, SpecMilestone.ELECTRA);
+    return maybeStateAndMetadata.map(
+        stateAndMetaData -> {
+          final SszList<PendingDeposit> deposits =
+              stateAndMetaData.getData().toVersionElectra().orElseThrow().getPendingDeposits();
+          return new ObjectAndMetaData<>(
+              deposits,
+              stateAndMetaData.getMilestone(),
+              stateAndMetaData.isExecutionOptimistic(),
+              stateAndMetaData.isCanonical(),
+              stateAndMetaData.isFinalized());
+        });
+  }
+
+  public SafeFuture<Optional<ObjectAndMetaData<SszList<PendingPartialWithdrawal>>>>
+      getPendingPartialWithdrawals(final String stateIdParam) {
+    return stateSelectorFactory
+        .createSelectorForStateId(stateIdParam)
+        .getState()
+        .thenApply(this::getPendingPartialWithdrawals);
+  }
+
+  private Optional<ObjectAndMetaData<SszList<PendingPartialWithdrawal>>>
+      getPendingPartialWithdrawals(final Optional<StateAndMetaData> maybeStateAndMetadata) {
+    checkMinimumMilestone(maybeStateAndMetadata, SpecMilestone.ELECTRA);
+
+    return maybeStateAndMetadata.map(
+        stateAndMetaData -> {
+          final SszList<PendingPartialWithdrawal> withdrawals =
+              stateAndMetaData
+                  .getData()
+                  .toVersionElectra()
+                  .orElseThrow()
+                  .getPendingPartialWithdrawals();
+          return new ObjectAndMetaData<>(
+              withdrawals,
+              stateAndMetaData.getMilestone(),
+              stateAndMetaData.isExecutionOptimistic(),
+              stateAndMetaData.isCanonical(),
+              stateAndMetaData.isFinalized());
+        });
+  }
+
+  private void checkMinimumMilestone(
+      final Optional<StateAndMetaData> maybeStateAndMetadata,
+      final SpecMilestone minimumMilestone) {
+    if (maybeStateAndMetadata.isPresent()
+        && maybeStateAndMetadata.get().getMilestone().isLessThan(minimumMilestone)) {
+      throw new BadRequestException(
+          String.format(
+              "The state was successfully retrieved, but was prior to %s and does not contain pending deposits.",
+              minimumMilestone.name()));
     }
-    return Optional.empty();
   }
 }